+++ src/Commands/StatusCommand.php
@@ -14,7 +14,7 @@ use Symfony\Component\Console\Style\SymfonyStyle;
 class StatusCommand extends Command
 {
 
-    use CommandTrait;
+    use \PHPUnuhi\Traits\CommandTrait;
 
     /**
      * @return void